requirements.txt
#####################
# conda create -n transdf python=3.10
# conda activate transdf

# we recommend using conda to install pytorch and torch-geometric dependencies
# conda install pytorch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0  pytorch-cuda=11.8 -c pytorch -c nvidia

# if failed, try to install pytorch and torch-geometric dependencies using pip
# pip install torch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0 --index-url https://download.pytorch.org/whl/cu118

# pip install https://data.pyg.org/whl/torch-2.4.0%2Bcu118/torch_cluster-1.6.3%2Bpt24cu118-cp310-cp310-linux_x86_64.whl
# pip install https://data.pyg.org/whl/torch-2.4.0%2Bcu118/torch_scatter-2.1.2%2Bpt24cu118-cp310-cp310-linux_x86_64.whl
# pip install https://data.pyg.org/whl/torch-2.4.0%2Bcu118/torch_sparse-0.6.18%2Bpt24cu118-cp310-cp310-linux_x86_64.whl
# pip install https://data.pyg.org/whl/torch-2.4.0%2Bcu118/torch_spline_conv-1.2.2%2Bpt24cu118-cp310-cp310-linux_x86_64.whl
# pip install -r requirements.txt
#####################

cuml-cu11
einops
ipykernel
isort
matplotlib
numba
openpyxl
pandas
patool
pot
pyrpca
pyyaml
PyWavelets
reformer-pytorch
robustica
scikit-learn
scikit-learn-intelex
scipy
seaborn
setproctitle
sktime
sympy
tensorboard
tensorboard-plugin-profile
thop
torchinfo
torch_dct
torch-geometric
torchmetrics
torch_tb_profiler
tqdm
tsdistances